+++ 3rdparty/exported/snmalloc/src/test/perf/singlethread/singlethread.cc
@@ -5,22 +5,21 @@
 
 using namespace snmalloc;
 
-template<ZeroMem zero_mem>
+template<typename Conts>
 void test_alloc_dealloc(size_t count, size_t size, bool write)
 {
-  auto& alloc = ThreadAlloc::get();
-
   {
     MeasureTime m;
     m << "Count: " << std::setw(6) << count << ", Size: " << std::setw(6)
-      << size << ", ZeroMem: " << (zero_mem == YesZero) << ", Write: " << write;
+      << size
+      << ", ZeroMem: " << std::is_same_v<Conts, Zero> << ", Write: " << write;
 
     std::unordered_set<void*> set;
 
     // alloc 1.5x objects
     for (size_t i = 0; i < ((count * 3) / 2); i++)
     {
-      void* p = alloc.alloc<zero_mem>(size);
+      void* p = snmalloc::alloc<Conts>(size);
       SNMALLOC_CHECK(set.find(p) == set.end());
 
       if (write)
@@ -36,13 +35,13 @@ void test_alloc_dealloc(size_t count, size_t size, bool write)
       void* p = *it;
       set.erase(it);
       SNMALLOC_CHECK(set.find(p) == set.end());
-      alloc.dealloc(p, size);
+      snmalloc::dealloc(p, size);
     }
 
     // alloc 1x objects
     for (size_t i = 0; i < count; i++)
     {
-      void* p = alloc.alloc<zero_mem>(size);
+      void* p = snmalloc::alloc<Conts>(size);
       SNMALLOC_CHECK(set.find(p) == set.end());
 
       if (write)
@@ -55,12 +54,12 @@ void test_alloc_dealloc(size_t count, size_t size, bool write)
     while (!set.empty())
     {
       auto it = set.begin();
-      alloc.dealloc(*it, size);
+      snmalloc::dealloc(*it, size);
       set.erase(it);
     }
   }
 
-  snmalloc::debug_check_empty<snmalloc::Alloc::Config>();
+  snmalloc::debug_check_empty();
 }
 
 int main(int, char**)
@@ -69,18 +68,18 @@ int main(int, char**)
 
   for (size_t size = 16; size <= 128; size <<= 1)
   {
-    test_alloc_dealloc<NoZero>(1 << 15, size, false);
-    test_alloc_dealloc<NoZero>(1 << 15, size, true);
-    test_alloc_dealloc<YesZero>(1 << 15, size, false);
-    test_alloc_dealloc<YesZero>(1 << 15, size, true);
+    test_alloc_dealloc<Uninit>(1 << 15, size, false);
+    test_alloc_dealloc<Uninit>(1 << 15, size, true);
+    test_alloc_dealloc<Zero>(1 << 15, size, false);
+    test_alloc_dealloc<Zero>(1 << 15, size, true);
   }
 
   for (size_t size = 1 << 12; size <= 1 << 17; size <<= 1)
   {
-    test_alloc_dealloc<NoZero>(1 << 10, size, false);
-    test_alloc_dealloc<NoZero>(1 << 10, size, true);
-    test_alloc_dealloc<YesZero>(1 << 10, size, false);
-    test_alloc_dealloc<YesZero>(1 << 10, size, true);
+    test_alloc_dealloc<Uninit>(1 << 10, size, false);
+    test_alloc_dealloc<Uninit>(1 << 10, size, true);
+    test_alloc_dealloc<Zero>(1 << 10, size, false);
+    test_alloc_dealloc<Zero>(1 << 10, size, true);
   }
 
   return 0;