+++ tests/client_protocols.py
@@ -135,7 +135,7 @@ def test_http2(network, args):
 
 def run(args):
     with infra.network.network(
-        args.nodes, args.binary_dir, args.debug_nodes, args.perf_nodes, pdb=args.pdb
+        args.nodes, args.binary_dir, args.debug_nodes, pdb=args.pdb
     ) as network:
         network.start_and_open(args)
         test_tls(network, args)
@@ -145,7 +145,7 @@ def run(args):
     args.http2 = True
     args.nodes = infra.e2e_args.nodes(args, 1)
     with infra.network.network(
-        args.nodes, args.binary_dir, args.debug_nodes, args.perf_nodes, pdb=args.pdb
+        args.nodes, args.binary_dir, args.debug_nodes, pdb=args.pdb
     ) as network:
         network.start_and_open(args)
         test_http2(network, args)